- Updated `target_fds` type to `uint32_t` for consistency with `fds_generation`. - Enhanced `/proc` and `/sys` pseudo-file read logic to ensure correctness of short reads, proper EOF handling, and type matching with `read()` return values. - Improved comments and code clarity on file descriptor and buffer management.

+++ src/collectors/apps.plugin/apps_pid_files.c
@@ -13,6 +13,9 @@ uint32_t all_files_len_get(void) {
 
 #if (PROCESSES_HAVE_FDS == 1)
 
+// Generation counter for tracking which fds have been aggregated in the current
+// iteration. target_fds[] entries are uint32_t to match this counter's type,
+// avoiding signed/unsigned conversion issues across the full uint32_t range.
 static uint32_t fds_generation = 0;
 
 void fds_generation_advance(void) {
@@ -61,8 +64,8 @@ static inline void reallocate_target_fds(struct target *w) {
         return;
 
     if(unlikely(!w->target_fds || w->target_fds_size < all_files_size)) {
-        w->target_fds = reallocz(w->target_fds, sizeof(int) * all_files_size);
-        memset(&w->target_fds[w->target_fds_size], 0, sizeof(int) * (all_files_size - w->target_fds_size));
+        w->target_fds = reallocz(w->target_fds, sizeof(uint32_t) * all_files_size);
+        memset(&w->target_fds[w->target_fds_size], 0, sizeof(uint32_t) * (all_files_size - w->target_fds_size));
         w->target_fds_size = all_files_size;
     }
 }
@@ -111,13 +114,13 @@ static inline void aggregate_fd_on_target(int fd, struct target *w) {
     if(unlikely(!w))
         return;
 
-    if(w->target_fds[fd] == (int)fds_generation) {
+    if(w->target_fds[fd] == fds_generation) {
         // it is already aggregated this iteration
         return;
     }
 
     // mark as seen this generation
-    w->target_fds[fd] = (int)fds_generation;
+    w->target_fds[fd] = fds_generation;
 
     aggregage_fd_type_on_openfds(all_files[fd].type, &w->openfds);
 }
